Add files via upload
I fixed the issue with Victor's CSS file not displaying properly. I also stripped out a lot of the excess code and reformatted the code to hopefully be easier to work with. Finally, the chess pieces are now able to move their position, but are not restricted to only legal moves.